+++ bl_update/main.c
@@ -39,9 +39,9 @@ static uint32_t input_pin;
 static void delayMicroseconds(uint32_t micros)
 {
     while (micros > 0) {
-        uint32_t us = micros>10000?10000:micros;
-        const uint32_t us_start = bl_timer_us();
-        while (bl_timer_us() - us_start < us) ;
+        uint16_t us = micros>10000?10000:micros;
+        const uint16_t us_start = bl_timer_us();
+        while ((uint16_t)(bl_timer_us() - us_start) < us) ;
         micros -= us;
     }
 }
